[#59] Additional cleanup TODOs completed.
38. [DONE] In USEC_SLEEP() change * 1000 to * 1000L so result is correct
    type. Also do the same in gt_timer.c line 348 or so.

39. [DONE] Change ydb_zmaxtptime to ydb_maxtptime (like TODO 27 said) and
    fix test.

40. [DONE] Remove YDB_ERR_UNKNOWN error.

42. [DONE] Change LVUNDEF_OK to ERR_LVUNDEF_OK.

43. [DONE] Change YDB_MAX_NAMES to 35.

44. [DONE] Fix error with ydb_gbldir not being picked up.

45. [DONE] Add INVNAMECOUNT check and error to ydb_tp_s() and
    ydb_delete_excl_s().

46. [DONE] All error checking including parameters should occur before
    all locks are released in ydb_lock_s().

47. [DONE] All ydb_* functions should return either int or void - except
    the one oddball ydb_malloc() which returns a void *.

48. [DONE] Rename YDB_ERROR_SEVERITY() macro to YDB_SEVERITY() and the
    severity codes to YDB_SEVERITY_*.

+++ sr_unix/gt_timers.c
@@ -345,7 +345,7 @@ void hiber_start(uint4 hiber)
 	{	/* normally, if SIGALRMs are blocked, we must already be inside a timer handler, but someone can actually disable
 		 * SIGALRMs, in which case we do not want this assert to trip in pro */
 		assert(1 <= timer_stack_count);
-		SLEEP_USEC(hiber * 1000, TRUE);
+		SLEEP_USEC((long)(hiber * 1000L), TRUE);
 	} else
 	{
 		assertpro(1 > timer_stack_count);	/* if SIGALRMs are not blocked, we cannot be inside a timer handler */
